+++ src/pages/changelog.md
@@ -18,6 +18,7 @@ This also includes the [Activity Timeline](#timeline), a visual representation o
 ## 3.19.8
 
 - Reviewed /likes archive and added annual grouping similar to the blog
+- Fix /reading (data from Raindrop.io)
 
 ## 3.19.7